Refactor: Reorganize tracking screens and add Map feature
- Move track_screen.dart to live_tracking/screens directory
- Remove old live_tracking_screen.dart
- Update home_screen.dart with new navigation structure
- Add new Map feature directory
- Maintain isTracking and hasActiveSession as separate properties for proper GPS and session state management
